// src/app/cards/enum.ts
export enum Diplome{
    BM = 'Bachelor + Master',
    PM = 'Prépa + Master',
    B = 'Bachelor',
    M = 'Master'
}

export enum DiplomeYear{
    B1 = 'Bachelor 1',
    B2 = 'Bachelor 2',
    B3 = 'Bachelor 3',
    M1 = 'Master 1',
    M2 = 'Master 2',
    P1 = 'Prépa 1',
    P2= 'Prépa 2',
}

export enum Faculty {
    HORDE = 'La Horde',
    TURING = 'Turing',
}

export enum Course{
    PROG = 'Programmation',
    ART = 'Game Art',
    GD = 'Game Design',
    TA = 'Tech Art',
    PROD = 'Producing',
    SOUND = 'Sound Design',
    BUSN = 'Business & Événementiel',
    DATA = 'IA & Big Data',
    LOGI = 'Génie Logiciel',
    CLOUD = 'Cloud & Cybersécurité',
}